Dithering

+++ src/sound/soundmacros.h
@@ -2,11 +2,11 @@
 #define SOUND_SOUNDMACROS_H
 
 #define SOUND_ENUMERATE_TYPE(a_macro, a_param) \
-    a_macro(Float32, a_param)                  \
-        a_macro(Float64, a_param)              \
-            a_macro(Int8, a_param)             \
-                a_macro(Int16, a_param)        \
-                    a_macro(Int32, a_param)
+    a_macro(Int8, a_param)                     \
+        a_macro(Int16, a_param)                \
+            a_macro(Float32, a_param)          \
+                a_macro(Int32, a_param)        \
+                    a_macro(Float64, a_param)
 
 #define SOUND_INSTANTIATE_METHOD_ONE(a_type, a_macro) \
     a_macro(a_type)
